Constant speed movement working; self-collision detecting present, but buggy
Still using the spring forces and RK4 integration from gafferongames.com,
with a spring "goal" (equilibrium) that is always a fixed length away from
the top of the snake.

The collision detection bugs me. Either I missunderstood something or there's
a bug in android.graphics.Region.setPath. It does reliably detect when the
head touches the tail of the path, but there are also false positives.